requirements.txt
discord.py[voice]
Pillow
httpx
opencv-python
numpy
yt-dlp
d20